requirements.txt
# Production-Ready Requirements for AI Interview v2.0

# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM 
groq>=0.4.0
tenacity>=8.2.0

# Speech & Audio
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pyttsx3>=2.90
# wave is built-in to Python, no need to install

# Computer Vision & Face Analysis
opencv-python>=4.8.0
mediapipe>=0.10.0
deepface>=0.0.79
tensorflow>=2.14.0  # Required by DeepFace
tf-keras>=2.14.0  # Required by TensorFlow 2.20+

# NLP & Text Processing
spacy>=3.6.0
textblob>=0.17.0
gensim>=4.3.0
nltk>=3.8.0

# Resume Parsing
PyPDF2>=3.0.0
python-docx>=1.0.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Reporting
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
alembic>=1.12.0  # Migrations

# Monitoring & Observability
psutil>=5.9.0  # System metrics
sentry-sdk>=1.38.0  # Error tracking (optional)

# Security
cryptography>=41.0.0

# Caching (optional but recommended)
redis>=5.0.0
hiredis>=2.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pylint>=2.17.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6

# Development
ipython>=8.14.0
streamlit>=1.28.0
nest-asyncio>=1.5.0


# Note: After installing, download spacy model:
# python -m spacy download en_core_web_sm
